src/auth/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
jinja2==3.1.2
python-multipart==0.0.6
pymongo==4.3.3
python-jose==3.3.0
passlib==1.7.4

